Fixes, documentation, and thread names

+++ src/Coordination/KeeperRequestDispatcher2.h
@@ -16,9 +16,40 @@ extern template class NonblockingBoundedQueue<DB::KeeperResponseForSession>;
 namespace DB
 {
 
-//asdqwe consider explaining the series of tubes mental model, with a toy pipeline
+/// Takes client requests from KeeperTCPHandler (or whoever else calls putRequest), passes them
+/// to nuraft leader, detects when they're committed (or failed), and passes responses back to
+/// KeeperTCPHandler (or whoever registers their response callback by calling registerSession).
 ///
-/// KeeperRequestDispatcher2 is in charge of preventing queue size bloat in the whole pipeline.
+/// Considerations:
+///  * Forwarding. If nuraft leader is not on this node, requests are sent to the leader node.
+///  * Batching. responseThread() picks as many consecutive requests as possible to send together,
+///    to save the overhead of passing tiny individual requests around.
+///  * Pipelining. Multiple request batches can be in flight simultaneously, to keep the whole
+///    nuraft pipeline busy, for full resource utilization for maximum throughput.
+///    E.g. maybe one batch is being processed by followers, while the next batch is being
+///    transferred from leader to followers, while the next batch is being preprocessed by leader.
+///  * Ordering. Requests from one client session must be executed in the order they arrived.
+///    To preserve order even when multiple requests are executed concurrently, we rely on nuraft
+///    leader not reordering requests, on tcp not reordering messages, and on all the plumbing in
+///    between also preserving order. See KeeperAppendStream.
+///    If leader changes or we lose connection to it, we fail all in-flight requests (their outcome
+///    is unknown) and open a new stream.
+///  * Executing reads locally. Read requests don't need to go through raft, and don't even have to
+///    be forwarded to the leader node. But they must be executed at exactly the right time: after
+///    all previous write requests in the same client session were applied, before the next
+///    request in the session is applied.
+///  * Reordering requests. Requests in the same client session can't be reordered, but we're free
+///    to interleave requests from different sessions however we like. We use this freedom to group
+///    read requests together as much as possible. When we have many consecutive read requests, we
+///    can execute them in parallel (while blocking writes, so that znode state doesn't change).
+///  * Minimizing load on the commit thread. Commit thread is usually the bottleneck for throughput
+///    of the whole keeper cluster. We should do as little as possible in the commit callback.
+///  * Flow control. See below.
+///
+/// KeeperRequestDispatcher2 is in charge of most of flow control in keeper server.
+/// I.e. preventing queue size bloat in the whole keeper server + nuraft pipeline. Especially if
+/// some part of the pipeline is not keeping up; e.g. clients are sending more requests than nuraft
+/// can process, or keeper server is trying to send more replies than clients are ready to receive.
 /// Here's a story for why no queues or buffers throughout keeper can grow out of control:
 ///  * Requests from clients arrive in server's TCP sockets (memory usage limited by the OS) and go
 ///    through some ReadBuffer-s (size limited per client connection), and are parsed by KeeperTCPHandler.
@@ -145,7 +176,7 @@ private:
     std::atomic<bool> current_stream_is_suspect {};
 
     NonblockingBoundedQueue<KeeperRequestForSession> requests_queue;
-    std::atomic<size_t> requests_queue_bytes {};
+    std::atomic<int64_t> requests_queue_bytes {};
 
     /// TODO: Currently responses to all requests are formed on all nodes, and go through this queue
     ///       and through responseThread and session id lookup on all nodes, and all except one node
@@ -153,7 +184,7 @@ private:
     ///       early that response is not needed. It would also save time in
     ///       KeeperRequestDispatcher2::onCommit, we won't have to check most requests against the queue.
     NonblockingBoundedQueue<KeeperResponseForSession> responses_queue;
-    std::atomic<size_t> response_bytes_in_all_queues {};
+    std::atomic<int64_t> response_bytes_in_all_queues {};
 
     /// State frequently mutated by dispatch thread.
     alignas(CH_CACHE_LINE_SIZE) std::atomic<size_t> tail_idx {};
